+++ properties/Connection/SelectWhereIn.php
@@ -1,0 +1,102 @@
+<?php
+declare(strict_types = 1);
+
+namespace Properties\Formal\AccessLayer\Connection;
+
+use Formal\AccessLayer\{
+    Query\SQL,
+    Query\Parameter,
+    Query\Select,
+    Table\Name,
+    Connection,
+};
+use Innmind\Specification\{
+    Comparator,
+    Composable,
+    Sign,
+};
+use Innmind\BlackBox\{
+    Property,
+    Set,
+    Runner\Assert,
+};
+
+/**
+ * @implements Property<Connection>
+ */
+final class SelectWhereIn implements Property
+{
+    private string $uuid;
+    private string $username;
+    private int $number;
+
+    public function __construct(string $uuid, string $username, int $number)
+    {
+        $this->uuid = $uuid;
+        $this->username = $username;
+        $this->number = $number;
+    }
+
+    public static function any(): Set
+    {
+        return Set\Composite::immutable(
+            static fn(...$args) => new self(...$args),
+            Set\Uuid::any(),
+            Set\Strings::madeOf(Set\Chars::ascii())->between(0, 255),
+            Set\Integers::any(),
+        );
+    }
+
+    public function applicableTo(object $connection): bool
+    {
+        return true;
+    }
+
+    public function ensureHeldBy(Assert $assert, object $connection): object
+    {
+        $insert = SQL::of('INSERT INTO `test` VALUES (?, ?, ?);')
+            ->with(Parameter::of($this->uuid))
+            ->with(Parameter::of($this->username))
+            ->with(Parameter::of($this->number));
+        $connection($insert);
+
+        $select = Select::from(new Name('test'));
+        $select = $select->where(new class($this->uuid) implements Comparator {
+            use Composable;
+
+            public function __construct(private string $uuid)
+            {
+            }
+
+            public function property(): string
+            {
+                return 'test.id';
+            }
+
+            public function sign(): Sign
+            {
+                return Sign::in;
+            }
+
+            public function value(): array
+            {
+                return [$this->uuid];
+            }
+        });
+        $rows = $connection($select);
+
+        $assert->count(1, $rows);
+        $assert->same(
+            $this->uuid,
+            $rows
+                ->first()
+                ->flatMap(static fn($row) => $row->column('id'))
+                ->match(
+                    static fn($id) => $id,
+                    static fn() => null,
+                ),
+        );
+
+        return $connection;
+    }
+}